+++ dspot/src/test/java/eu/stamp_project/dspot/assertiongenerator/AssertionRemoverTest.java
@@ -3,7 +3,7 @@ package eu.stamp_project.dspot.assertiongenerator;
 import eu.stamp_project.AbstractTest;
 import eu.stamp_project.Utils;
 import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.AssertionRemover;
-import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.testmethodreconstructor_components.observer_components.TestWithLogGenerator;
+import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.methodreconstructor_components.observer_components.TestWithLogGenerator;
 import eu.stamp_project.test_framework.TestFramework;
 import eu.stamp_project.utils.AmplificationHelper;
 import org.junit.BeforeClass;
