+++ dspot/src/test/java/eu/stamp_project/dspot/assertiongenerator/TranslatorTest.java
@@ -2,7 +2,7 @@ package eu.stamp_project.dspot.assertiongenerator;
 
 import eu.stamp_project.AbstractTest;
 import eu.stamp_project.Utils;
-import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.testmethodreconstructor_components.assertionsyntaxbuilder_components.Translator;
+import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.methodreconstructor_components.assertionsyntaxbuilder_components.Translator;
 import org.jetbrains.annotations.NotNull;
 import org.junit.Before;
 import org.junit.Test;
