+++ dspot/src/test/java/eu/stamp_project/dspot/assertiongenerator/performancetest/performancetest_components/MethodReconstructorWithTime.java
@@ -1,16 +1,16 @@
-package eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components;
+package eu.stamp_project.dspot.assertiongenerator.performancetest.performancetest_components;
 
 import eu.stamp_project.compare.Observation;
 import eu.stamp_project.dspot.AmplificationException;
-import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.testmethodreconstructor_components.AssertionSyntaxBuilder;
-import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.testmethodreconstructor_components.Observer;
+import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.MethodReconstructor;
+import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.methodreconstructor_components.AssertionSyntaxBuilder;
 import eu.stamp_project.dspot.assertiongenerator.utils.AssertionGeneratorUtils;
+import eu.stamp_project.utils.program.InputConfiguration;
 import eu.stamp_project.utils.AmplificationHelper;
 import eu.stamp_project.utils.CloneHelper;
 import eu.stamp_project.utils.Counter;
 import eu.stamp_project.utils.DSpotUtils;
 import eu.stamp_project.utils.compilation.DSpotCompiler;
-import eu.stamp_project.utils.program.InputConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import spoon.reflect.code.CtBlock;
@@ -26,41 +26,38 @@ import spoon.reflect.visitor.filter.TypeFilter;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 /**
  * Created by Benjamin DANGLOT
  * benjamin.danglot@inria.fr
- * on 3/3/17
+ * on 18/07/18
  */
-public class TestMethodReconstructor {
+public class MethodReconstructorWithTime extends MethodReconstructor {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(TestMethodReconstructor.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(MethodReconstructor.class);
 
     private Factory factory;
 
     private InputConfiguration configuration;
 
-    private Observer observer;
+    private ObserverWithTime observerWithTime;
 
-    public TestMethodReconstructor(CtType originalClass,
-                                   InputConfiguration configuration,
-                                   DSpotCompiler compiler,
-                                   Map<CtMethod<?>, List<CtLocalVariable<?>>> variableReadsAsserted) {
+    public MethodReconstructorWithTime(CtType originalClass, InputConfiguration configuration, DSpotCompiler compiler, Map<CtMethod<?>, List<CtLocalVariable<?>>> variableReadsAsserted) {
+        super(originalClass, configuration, compiler, variableReadsAsserted);
         this.configuration = configuration;
         this.factory = configuration.getFactory();
-        this.observer = new Observer(originalClass,
+        this.observerWithTime = new ObserverWithTime(originalClass,
                 configuration,
                 compiler,
                 variableReadsAsserted);
     }
 
     /**
-     * Adds new assertions in multiple passing tests.
+     * Adds new assertions in multiple tests.
      * <p>
      * <p>Instruments the tests to have observation points.
-     * Details in {@link Observer#getObservations(CtType, List)}.
+     * Details in {@link ObserverWithTime#getObservations(CtType, List)}.
      * <p>
      * <p>Details of the assertion generation in {@link #buildTestWithAssert(CtMethod, Map)}.
      *
@@ -71,12 +68,26 @@ public class TestMethodReconstructor {
     public List<CtMethod<?>> addAssertions(CtType<?> testClass, List<CtMethod<?>> testCases) {
         Map<String, Observation> observations;
         try {
-            observations = observer.getObservations(testClass, testCases);
+            observations = observerWithTime.getObservations(testClass, testCases);
         } catch (AmplificationException e) {
             e.printStackTrace();
             return Collections.emptyList();
         }
-        return buildEachTest(testCases,observations);
+        this.timeInstrumentation = observerWithTime.timeInstrumentation;
+        this.timeRunningInstrumentation = observerWithTime.timeRunningInstrumentation;
+        final long start = System.currentTimeMillis();
+        final List<CtMethod<?>> generation = buildEachTest(testCases,observations);
+        this.timeGeneration = System.currentTimeMillis() - start;
+        return generation;
+    }
+
+    public long timeInstrumentation;
+    public long timeRunningInstrumentation;
+    public long timeGeneration;
+
+    public void reset() {
+        this.observerWithTime.reset();
+        this.timeGeneration = 0;
     }
 
     // add assertions to each test with values retrieved from logs
@@ -89,8 +100,7 @@ public class TestMethodReconstructor {
 
     /**
      * Adds new assertions to a test from observation points.
-     * <p>
-     * <p>Details of constructing the syntax for the assertions in {@link AssertionSyntaxBuilder#buildAssert(CtMethod, Set, Map, Double)}.
+     *
      * @param test         Test method
      * @param observations Observation points of the test suite
      * @return Test with new assertions
@@ -128,7 +138,7 @@ public class TestMethodReconstructor {
     }
 
     private void goThroughAssertionStatements(List<CtStatement> assertStatements,String id,
-                                               List<CtStatement> statements,Integer numberOfAddedAssertion){
+                                              List<CtStatement> statements,Integer numberOfAddedAssertion){
         int line = Integer.parseInt(id.split("__")[1]);
         CtStatement lastStmt = null;
         for (CtStatement assertStatement : assertStatements) {
@@ -153,7 +163,7 @@ public class TestMethodReconstructor {
     }
 
     private void decideInvocationReplacement(CtStatement statementToBeAsserted,String id,CtStatement assertStatement,
-                                   List<CtStatement> statements,int line,CtStatement lastStmt){
+                                             List<CtStatement> statements,int line,CtStatement lastStmt){
 
         /* if the statement to be asserted is a method or constructor call, replace that invocation in the
         assertion with an equivalent local variable */
@@ -174,7 +184,7 @@ public class TestMethodReconstructor {
         // create a new local variable and assign the invocation to it
         CtInvocation invocationToBeReplaced = (CtInvocation) statementToBeAsserted.clone();
         final CtLocalVariable localVariable = factory.createLocalVariable(
-                AssertionGeneratorUtils.getCorrectTypeOfInvocation(invocationToBeReplaced),
+                invocationToBeReplaced.getType(),
                 "o_" + id.split("___")[0],
                 invocationToBeReplaced
         );
